refactor: consolidate footer composable
