+++ apps/web/app/components/blocks/Footer/FooterForm.vue
@@ -349,6 +349,7 @@ const { data } = useCategoryTemplate(
 );
 const { blockUuid } = useSiteConfiguration();
 const { findOrDeleteBlockByUuid } = useBlockManager();
+const { mapFooterData, FOOTER_SWITCH_DEFINITIONS } = useFooter();
 const props = defineProps<{ uuid?: string }>();
 
 const firstColumnOpen = ref(false);
